+++ Runtime/World/Components/Camera.h
@@ -50,7 +50,7 @@ namespace Spartan
     class SPARTAN_CLASS Camera : public IComponent
     {
     public:
-        Camera(Context* context, Entity* entity, uint32_t id = 0);
+        Camera(Context* context, Entity* entity, uint64_t id = 0);
         ~Camera() = default;
 
         //= ICOMPONENT ===============================
