+++ Runtime/World/Components/Light.h
@@ -63,7 +63,7 @@ namespace Spartan
     class SPARTAN_CLASS Light : public IComponent
     {
     public:
-        Light(Context* context, Entity* entity, uint32_t id = 0);
+        Light(Context* context, Entity* entity, uint64_t id = 0);
         ~Light() = default;
 
         //= COMPONENT ================================
